Continued QuadTree Updates
Work on making QuadTree work for both "game" objects and HTML elements.
Developers will have to enter in target size and position in the future.
This is not a working implementation yet, but targets will be stored as
an object as thus:

item: The target inserted into the QuadTree
width: Width of the target
height: of the target
x: x position of the target
y: y position of the target